[refs #460] Fix intermittent faint strikethrough rendering on IE11/Edge
As mentioned in the patch, the border width increase does not change the
rendering of the strike in Win/Mac versions of Chrome or Firefox, nor in
desktop Safari or iOS Safari (bumping to 1.5, however,_does_ result in a
noticeable thickening, because browsers round this up to 2px), but is
enough to kick IE/Edge into rendering a visible 1px line all of the time.
The downside to this approach is that it doesn't seem like the rendering
of fractional pixel widths for borders is defined in the CSS spec, so
future updates to browsers could potentially render this differently.